commit before rewrite

+++ 02_Kartenspiel/src/de/esg/ausbildung/honl/game/SaveData.java
@@ -2,5 +2,5 @@ package de.esg.ausbildung.honl.game;
 
 import java.util.ArrayList;
 
-public record SaveData(double balance, int[] score, ArrayList<Card> cardStack) { 
+public record SaveData(int balance, int[] score, ArrayList<Card> cardStack) {
 }
